glue-codes/kitefast/preprocessor/requirements.txt
numpy>=1.16.1
PyYAML>=3.12
